requirements.txt
altgraph==0.17.4
brainpy==2.4.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.49.0
jax==0.4.12
jaxlib==0.4.12
kiwisolver==1.4.5
llvmlite==0.42.0
matplotlib==3.8.3
ml-dtypes==0.3.2
msgpack==1.0.8
numba==0.59.0
numpy==1.24.3
opt-einsum==3.3.0
packaging==24.0
pillow==10.2.0
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
scipy==1.12.0
six==1.16.0
tqdm==4.66.2
